Add `variable.LengthLimited` schema.
